Rust API: Switch from f64 to f32 throughout protocol and client
The wire format maxes out at f32 or scaled integers, so f64 provided
no precision benefit. Using f32 halves the size of command/query
structs and is more appropriate for the no_std protocol crate on
embedded platforms where f64 may require software emulation.

+++ lib/rust/moteus/src/controller.rs
@@ -371,7 +371,7 @@ impl Controller {
     /// Builds a set-output-nearest command.
     ///
     /// Sets the output position to the nearest value matching the given position.
-    pub fn make_set_output_nearest(&self, position: f64, query: bool) -> Command {
+    pub fn make_set_output_nearest(&self, position: f32, query: bool) -> Command {
         let mut cmd = self.prepare_command(query);
         OutputNearestCommand::new(position).serialize(cmd.frame_mut());
         if query {
@@ -383,7 +383,7 @@ impl Controller {
     /// Builds a set-output-exact command.
     ///
     /// Sets the output position to exactly the given value.
-    pub fn make_set_output_exact(&self, position: f64, query: bool) -> Command {
+    pub fn make_set_output_exact(&self, position: f32, query: bool) -> Command {
         let mut cmd = self.prepare_command(query);
         OutputExactCommand::new(position).serialize(cmd.frame_mut());
         if query {
